feat: remove improvement

+++ js/app.js
@@ -7,22 +7,13 @@ var app = new Vue({
     el: '#app', data: {
         sites: [{name: "youtube", url: "youtube.com", desc: "video hosting", bind: "yt"}, {
             name: "github", url: "github.com", desc: "hosting of IT projects", bind: "gh"
-        }],
-        search_line: '',
-        engines: {
+        }], search_line: '', engines: {
             ya: "https://yandex.ru/search/?text=",
             google: "https://www.google.com/search?ie=UTF-8&q=",
             brave: "https://search.brave.com/search?q=",
             ddg: "https://duckduckgo.com/?q=",
             searx: "https://searx.tiekoetter.com/search?q="
-        },
-        engines_show: [],
-        result: [],
-        add: false,
-        to_remove: false,
-        remove_index: -1,
-        engine: 'ya',
-        engine_request: []
+        }, engines_show: [], result: [], add: false, to_remove: false, remove_index: 0, engine: 'ya', engine_request: []
     }, methods: {
         search() {
             if (this.search_line[0] !== ':' && this.search_line[0] !== '/') {
@@ -37,9 +28,9 @@ var app = new Vue({
                     this.search_line = '';
                     this.add = false;
                     this.result = [];
-
                 }
-                if (this.to_remove && this.result.length === 1) {
+
+                if (this.to_remove && this.result.length > 0) {
                     this.sites.splice(this.remove_index, 1)
                     localStorage.setItem('sites', JSON.stringify(this.sites))
                     this.remove_index = -1;
@@ -47,6 +38,7 @@ var app = new Vue({
                     this.to_remove = false;
                     this.result = [];
                 }
+
                 if (this.engine_request.length) {
                     this.engines[this.engine_request[0]] = this.engine_request[1]
                     localStorage.setItem('engines', JSON.stringify(this.engines))
@@ -54,13 +46,12 @@ var app = new Vue({
                     this.engine_request = [];
                     this.search_line = '';
                 }
-                if (this.engines_show.length === 1 && this.to_remove) {
+                if (this.engines_show.length > 0 && this.to_remove) {
                     delete this.engines[this.engines_show[0]]
                     localStorage.setItem('engines', JSON.stringify(this.engines))
-
+                    console.log(this.engines_show[0])
                     let keys = Object.keys(this.engines);
                     if (localStorage.getItem('engine') === this.engines_show[0]) localStorage.setItem('engine', keys[keys.length * Math.random() << 0])
-
                     this.engines_show = [];
                     this.search_line = '';
                     this.to_remove = false;
@@ -139,19 +130,39 @@ var app = new Vue({
             let desc = this.search_line.slice(1).split(' ')[4];
 
             if (!isEmpty(name) && !isEmpty(url) && !isEmpty(bind)) {
+
+                for (let i in this.sites) {
+                    if (this.sites[i]['name'] === name) {
+                        return;
+                    }
+                }
+
                 this.add = {name: name, url: url, desc: desc, bind: bind};
             } else this.add = false;
 
         }, scmd_r(argument) {
             this.engines_show = []
             this.to_remove = false;
+            this.remove_index = 0;
             if (argument === 'e' || argument === 'engine') {
                 let name = this.search_line.slice(1).split(' ')[2];
                 if (!isEmpty(name)) {
                     for (let key in this.engines) {
                         if (key.includes(name)) {
+
                             this.engines_show.push(key);
-                            this.to_remove = true;
+                            // this.to_remove = true;
+
+                            if (this.engines_show.length === 1) {
+                                // this.remove_index = key;
+                                this.to_remove = true;
+                            }
+
+                            if (argument === this.engines[key]['name']) {
+                                // this.remove_index = key;
+                                this.to_remove = true;
+                            }
+
                         }
                     }
                 }
@@ -163,13 +174,23 @@ var app = new Vue({
             }
             for (let key in this.sites) {
                 if (this.sites[key]['name'].includes(argument)) {
+
                     this.result.push(this.sites[key]);
-                    this.remove_index = key;
-                    this.to_remove = true;
+
+                    if (this.result.length === 1) {
+                        this.remove_index = key;
+                        this.to_remove = true;
+                    }
+
+                    if (argument === this.sites[key]['name']) {
+                        this.remove_index = key;
+                        this.to_remove = true;
+                    }
                 }
             }
         }
     }, mounted() {
+        document.getElementById('app').style = "display: flex";
         if (!localStorage.getItem('engine')) {
             localStorage.setItem('engine', this.engine)
         }
